+++ src/components/FeaturedPost.js
@@ -1,5 +1,4 @@
 import * as React from 'react';
-
 import PropTypes from 'prop-types';
 import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
